SQL型の修正 serial→bigserial、int→bigint

+++ ddl/task_category.sql
@@ -1,6 +1,6 @@
 CREATE TABLE TASK_CATEGORY(
-  task_id int not null references TASK(id) on delete cascade,
-  category_id int not null references CATEGORY(id) on delete cascade,
+  task_id bigint not null references TASK(id) on delete cascade,
+  category_id bigint not null references CATEGORY(id) on delete cascade,
   created_at timestamp not null default current_timestamp,
   updated_at timestamp not null default current_timestamp,
   primary key (task_id, category_id)
