+++ ddl/task_tag.sql
@@ -1,5 +1,5 @@
 CREATE TABLE TASK_TAG (
-  task_id int not null references TASK(id) on delete cascade,
+  task_id bigint not null references TASK(id) on delete cascade,
   tag_id int not null references TAG(id) on delete cascade,
   created_at timestamp not null default current_timestamp,
   updated_at timestamp not null default current_timestamp,
